app/requirements.txt
click>=8.1.8
httpx>=0.28.1
langchain-google-genai>=2.0.10
langgraph>=0.3.18
langchain-openai>=0.1.0
pydantic>=2.10.6
python-dotenv>=1.1.0
uvicorn>=0.34.2
sse-starlette>=2.3.6
starlette>=0.46.2
a2a-sdk>=0.3.0
pyjwt>=2.10.1
cryptography>=44.0.2

sap-ai-sdk-gen[all]
